Convert phone number to long

+++ BookingState.py
@@ -40,7 +40,7 @@ class BookingState:
             ctr += 1
             if dbclub[j][day] != "" and ctr == self.state["club"]:
                 clubName = j
-        name = dbcont[self.state["originator"]]
+        name = dbcont[long(self.state["originator"])]
         payload = {
             "clubName": clubName,
             "name": name,
